# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfirmTheoryPlanRequest": {
        "properties": {
          "expected_decision_set_version": {
            "minimum": 1.0,
            "title": "Expected Decision Set Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_decision_set_version"
        ],
        "title": "ConfirmTheoryPlanRequest",
        "type": "object"
      },
...
      "ConfirmedTheoryPlanResponse": {
        "properties": {
          "adopted_candidate_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Adopted Candidate Ids",
            "type": "array"
          },
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/TheoryPlanAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "confirmed_at": {
            "format": "date-time",
            "title": "Confirmed At",
            "type": "string"
          },
          "confirmed_phenomenon": {
            "$ref": "#/components/schemas/PhenomenonSnapshotResponse"
          },
          "decision_set_id": {
            "format": "uuid",
            "title": "Decision Set Id",
            "type": "string"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/TheoryDecisionRecordResponse"
            },
            "title": "Decisions",
            "type": "array"
          },
          "knowledge_release_id": {
            "title": "Knowledge Release Id",
            "type": "string"
          },
          "match_run_id": {
            "format": "uuid",
            "title": "Match Run Id",
            "type": "string"
          },
          "phenomenon_query_id": {
            "format": "uuid",
            "title": "Phenomenon Query Id",
            "type": "string"
          },
          "phenomenon_version": {
            "title": "Phenomenon Version",
            "type": "integer"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/TheoryRelationResponse"
            },
            "title": "Relations",
            "type": "array"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "theory_plan_id": {
            "format": "uuid",
            "title": "Theory Plan Id",
            "type": "string"
          },
          "use_assignments": {
            "items": {
              "$ref": "#/components/schemas/TheoryUseAssignmentResponse"
            },
            "title": "Use Assignments",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "theory_plan_id",
          "task_id",
          "match_run_id",
          "decision_set_id",
          "version",
          "allowed_actions",
          "phenomenon_query_id",
          "phenomenon_version",
          "knowledge_release_id",
          "adopted_candidate_ids",
          "confirmed_phenomenon",
          "decisions",
          "use_assignments",
          "relations",
          "confirmed_at"
        ],
        "title": "ConfirmedTheoryPlanResponse",
        "type": "object"
      },
...
      "CreateResearchTaskRequest": {
        "properties": {
          "entry_type": {
            "$ref": "#/components/schemas/EntryType",
            "default": "direct_input"
          },
          "seed_theory_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Theory Id"
          }
        },
        "title": "CreateResearchTaskRequest",
        "type": "object"
      },
...
      "DeleteResearchTaskResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/ResearchTaskAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "deleted": {
            "const": true,
            "title": "Deleted",
            "type": "boolean"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "version",
          "allowed_actions",
          "deleted"
        ],
        "title": "DeleteResearchTaskResponse",
        "type": "object"
      },
...
      "EntryType": {
        "enum": [
          "direct_input",
          "material_input"
        ],
        "title": "EntryType",
        "type": "string"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MarkdownExportResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/ResearchTaskAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "contract_version": {
            "title": "Contract Version",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "media_type": {
            "const": "text/markdown",
            "title": "Media Type",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "version",
          "allowed_actions",
          "filename",
          "media_type",
          "markdown",
          "contract_version"
        ],
        "title": "MarkdownExportResponse",
        "type": "object"
      },
...
      "ResearchTaskPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResearchTaskNavigationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ResearchTaskPageResponse",
        "type": "object"
      },
...
      "ResearchTaskResponse": {
        "properties": {
          "allowed_actions": {
            "items": {
              "$ref": "#/components/schemas/ResearchTaskAction"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entry_type": {
            "$ref": "#/components/schemas/EntryType"
          },
          "seed_theory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Theory Id"
          },
          "seed_theory_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed Theory Name"
          },
          "status": {
            "$ref": "#/components/schemas/ResearchTaskStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "entry_type",
          "status",
          "version",
          "allowed_actions",
          "seed_theory_id",
          "seed_theory_name",
          "created_at",
          "updated_at"
        ],
        "title": "ResearchTaskResponse",
        "type": "object"
      },
...
      "TheoryPlanAction": {
        "enum": [
          "create_framework"
        ],
        "title": "TheoryPlanAction",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "群学致知前后端架构基线 API。",
    "title": "群学致知 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/decision-sets/{decision_set_id}/confirm": {
      "post": {
        "operationId": "confirm_theory_plan",
        "parameters": [
          {
            "in": "path",
            "name": "decision_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Decision Set Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmTheoryPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmedTheoryPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Confirm Theory Plan",
        "tags": [
          "matching"
        ]
      }
    },
...
    "/api/research-tasks": {
      "get": {
        "operationId": "list_research_tasks",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTaskPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "List Research Tasks",
        "tags": [
          "research-tasks"
        ]
      },
      "post": {
        "operationId": "create_research_task",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Research Task",
        "tags": [
          "research-tasks"
        ]
      }
    },
    "/api/research-tasks/{task_id}": {
      "delete": {
        "operationId": "delete_research_task",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResearchTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Delete Research Task",
        "tags": [
          "research-tasks"
        ]
      },
      "get": {
        "operationId": "get_research_task",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Research Task",
        "tags": [
          "research-tasks"
        ]
      }
    },
    "/api/research-tasks/{task_id}/export": {
      "get": {
        "operationId": "export_research_trace",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkdownExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Export Research Trace",
        "tags": [
          "research-tasks"
        ]
      }
    },
    "/api/research-tasks/{task_id}/frameworks": {
      "post": {
        "deprecated": true,
        "operationId": "create_framework",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFrameworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameworkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gone"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Create Framework",
        "tags": [
          "frameworks"
        ]
      }
    },
...
    "/api/research-tasks/{task_id}/navigation": {
      "get": {
        "operationId": "get_research_task_navigation",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTaskNavigationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Research Task Navigation",
        "tags": [
          "research-tasks"
        ]
      }
    },
...
    "/api/research-tasks/{task_id}/trace": {
      "get": {
        "operationId": "get_research_trace",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchTraceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Research Trace",
        "tags": [
          "research-tasks"
        ]
      }
    },
...
  }
}